setup api gateway
there is still an issue withe the io thouhg.

info comes in as a string insde the body key of the evnt.  so the preivous system doesn't work.
Also the body of the output needs to be converted into a string before it goes out.

so the issue with the decimal comes back again and needs to be solved.