can now deploy the client with one click
also made a script that runs tests on the server code before actually uploading it.
there's probably a python framework that would help me be a bit lazier that i don't know about.
Also need to remvoe the v1 cdk library i accidentally included in the project.  it was not necessary.